# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Indonesian Administrative Regions (Provinces, Cities, Districts, Villages)",
        "title": "Geo-ID API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cities/{id}": {
            "get": {
                "description": "Get specific city/regency details by its code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cities"
                ],
                "summary": "Get city by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City Code (e.g. 11.01)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Region"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cities/{id}/districts": {
            "get": {
                "description": "Get list of districts (Kecamatan) in a specific city",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cities"
                ],
                "summary": "Get districts in city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City Code (e.g. 11.01)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Region"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/districts/{id}": {
            "get": {
                "description": "Get specific district details by its code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "districts"
                ],
                "summary": "Get district by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "District Code (e.g. 11.01.01)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Region"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/districts/{id}/villages": {
            "get": {
                "description": "Get list of villages (Kelurahan/Desa) in a specific district",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "districts"
                ],
                "summary": "Get villages in district",
                "parameters": [
                    {
                        "type": "string",
                        "description": "District Code (e.g. 11.01.01)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Region"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/states": {
            "get": {
                "description": "Get list of all provinces in Indonesia",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get all states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Region"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/states/{id}": {
            "get": {
                "description": "Get specific province details by its code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get state by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State Code (e.g. 11)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Region"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/states/{id}/cities": {
            "get": {
                "description": "Get list of cities/regencies in a specific province",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get cities in state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State Code (e.g. 11)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Region"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/villages/{id}": {
            "get": {
                "description": "Get specific village details by its code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villages"
                ],
                "summary": "Get village by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Village Code (e.g. 11.01.01.2001)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Region"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Region": {
            "description": "Region information",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}